PERA-2832 :: Fix status bar bg color issue for testnet
